+++ CHANGELOG.md
@@ -150,6 +150,8 @@
 127. Fixed Issue with ALPHA Setting in Installer not abiding by Selected Theme.
 128. Fixed Issue with MelonPreferences Auto-Downgrade System in Installer when clean installing.
 129. Fixed Issue with Assembly Generator not Logging Launch Arguments.
+130. Added LaunchOptions.ini to allow for Configuration of Launch Options without using the Command Line.
+131. Fixed Issue where Launch Option "--melonloader.hidewarnings" would not function correctly.
 
 ---